// src/types.ts
export interface SharingCoefficient {
	tokenID: string;
	coefficient: Buffer;
}

export interface ValidatorEntry {
	address: string;
	name: string;
	blsKey: string;
	proofOfPossession: string;
	generatorKey: string;
	lastGeneratedHeight: number;
	isBanned: boolean;
	reportMisbehaviorHeights: number[];
	consecutiveMissedBlocks: number;
	commission: number;
	lastCommissionIncreaseHeight: number;
	sharingCoefficients: StakeSharingCoefficient[];
}

export interface ValidatorEntryBuffer
	extends Omit<ValidatorEntry, 'address' | 'sharingCoefficients'> {
	address: Buffer;
	sharingCoefficients: StakeSharingCoefficient[];
}

export interface Stake {
	validatorAddress: string;
	amount: bigint;
	sharingCoefficients: SharingCoefficient[];
}

export interface StakeBuffer {
	validatorAddress: Buffer;
	amount: bigint;
	sharingCoefficients: StakeSharingCoefficient[];
}

export interface PendingUnlockBuffer {
	validatorAddress: Buffer;
	amount: bigint;
	unstakeHeight: number;
}

export interface Staker {
	address: string;
	stakes: Stake[];
	pendingUnlocks: {
		validatorAddress: string;
		amount: bigint;
		unstakeHeight: number;
	}[];
}

export interface StakerBuffer extends Omit<Staker, 'address' | 'stakes' | 'pendingUnlocks'> {
	address: Buffer;
	stakes: StakeBuffer[];
	pendingUnlocks: PendingUnlockBuffer[];
}

export interface GenesisDataEntry {
	initRounds: number;
	initValidators: string[];
}

export interface PoSStoreEntry {
	validators: ValidatorEntry[];
	stakers: Staker[];
	genesisData: GenesisDataEntry;
}

export interface GenesisAssetEntry {
	module: string;
	data: Record<string, unknown>;
	schema: Schema;
}

export interface DelegateWeight {
	readonly address: Buffer;
	readonly voteWeight: bigint;
}
// ...
export interface NetworkConfigLocal {
	name: string;
	tokenID: string;
	prevSnapshotBlockHeight: number;
	additionalAccounts: { address: Buffer; balance: bigint }[];
}

export interface ForgingStatus {
	readonly address: string;
	readonly enabled: boolean;
	readonly height: number;
	readonly maxHeightPrevoted: number;
	readonly maxHeightGenerated: number;
}

export interface FileInfo {
	readonly fileName: string;
	readonly fileDir: string;
	readonly filePath: string;
}

export type BLSTransaction = {
	senderAddress: Buffer;
	params: {
		blsKey: Buffer;
		generatorKey: Buffer;
		proofOfPossession: Buffer;
	};
};